Changed double to float and removed permission laready set error

+++ inc/control_iface.h
@@ -114,7 +114,7 @@ int8_t npnt_reset_handle(npnt_s *handle);
 
 int8_t npnt_verify_permart(npnt_s *handle);
 
-int8_t npnt_alloc_and_get_fence_points(npnt_s* handle, float** vertx, float** verty);
+int8_t npnt_alloc_and_get_fence_points(npnt_s* handle, double** vertx, double** verty);
 
 int8_t npnt_get_max_altitude(npnt_s* handle, float* altitude);
